factor out contrib dependencies to simplify core installation
