+++ openssa/contrib/streamlit_ssa_prob_solver/requirements.txt
@@ -1,1 +1,1 @@
-OpenSSA
+OpenSSA[contrib]
