+++ openssa/core/ooda_rag/solver.py
@@ -1,7 +1,7 @@
+from openssa.core.ooda_rag.custom import CustomSSM
 from openssa.core.ooda_rag.ooda_rag import Solver
 from openssa.core.ooda_rag.heuristic import DefaultOODAHeuristic
 from openssa.core.ooda_rag.tools import ReasearchAgentTool
-from openssa.contrib.custom import CustomSSM
 from openssa.utils.llm_config import LLMConfig
 from openssa.utils.aitomatic_llm_config import AitomaticLLMConfig
