+++ pyproject.toml
@@ -69,9 +69,6 @@ fsspec = ">=2023.10,<2023.12"  # base/local file system
 adlfs = ">=2023.10,<2023.12"  # Azure Data Lake
 gcsfs = ">=2023.10,<2023.12"  # Google Cloud Storage
 s3fs = ">=2023.10,<2023.12"  # S3
-# contrib-supporting dependencies
-streamlit = ">=1.29"
-streamlit-mic-recorder = ">=0.0.4"
 # misc / other
 click = ">=8.1"
 pydantic = ">=2.5"
@@ -84,6 +81,8 @@ pylint = {version = ">=3.0", optional = true}
 pytest = {version = ">=7.4", optional = true}
 pillow = {version = ">=9.5", optional = true}
 python-pptx = {version = ">=0.6", optional = true}
+streamlit = {version = ">=1.29", optional = true}
+streamlit-mic-recorder = {version = ">=0.0.4", optional = true}
 torch = {version = ">=2.0", optional = true}
 transformers = {version = ">=4.35", optional = true}
 
@@ -93,6 +92,10 @@ lint = [
 ]
 test = [
   "pytest",
+]
+contrib = [
+  "streamlit",
+  "streamlit-mic-recorder",
 ]
 cv = [  # for processing images either standalone or in files such as .PPTX
   "pillow",
